pyproject.toml
[project]
name = "key4"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrey Tikhonov",email = "andrey.tikhonov.1974@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "yfinance (>=0.2.61,<0.3.0)",
    "lxml (>=5.4.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.loader.dependencies]
asyncpg = "^0.30.0"
aiofiles = "^24.1.0"
python-dotenv = "^1.1.0"

